requirements.txt
moviepy==1.0.3
numpy==1.26.4
Pillow==10.4.0
scikit-learn==1.5.0
TTS==0.22.0
